renamed simulation_parameters.py in simulation_scenario.py and moved it to starling_sim/
