--- starling_sim/basemodel/parameters/__init__.py
@@ -1,3 +1,0 @@
-"""
-This package contains the modules related to the parameter's management
-"""
